fix(library): infinite loop + directory false-positives for uppercase/extension-named content
getMediaParseStem infinite loop on uppercase extensions:
- path.basename is case-sensitive on Linux; basename('movie.MKV', '.mkv') returns 'movie.MKV' unchanged, so the while loop never terminates.
  - Any file with an uppercase extension (.MKV, .MP4, etc.) causes 100% CPU and a completely frozen event loop until the container is restarted.; so use the original-case extname() result to call basename(), only lowercase it for the TRANSPORT_AND_MEDIA_EXTENSIONS set lookup.

Scanner descends into video-extension directories:
- Directories named like video files (e.g. "matrix.mkv/" created by video editing software) pass all existing shouldExcludeFolder checks. The scanner enters them, adds every render/export as unmatched media, and the watcher triggers repeated rescans as new files appear; shouldExcludeFolderLegacy now skips any directory whose name is recognised by isVideoFile() (has a known video file extension).

Refactor:
 - move getMediaParseStem to shared media-utils module
