+++ src/lib/server/library/manual-import-service.ts
@@ -32,11 +32,8 @@ import {
 } from '$lib/server/library/LibraryAddService.js';
 import { getLibraryEntityService } from '$lib/server/library/LibraryEntityService.js';
 import { isLikelyAnimeMedia } from '$lib/shared/anime-classification.js';
-import {
-	extractSeasonFromPath,
-	getMediaParseStem,
-	resolveTvEpisodeIdentifier
-} from '$lib/server/library/tv-episode-resolver.js';
+import { getMediaParseStem } from '$lib/server/library/media-utils.js';
+import { extractSeasonFromPath, resolveTvEpisodeIdentifier } from '$lib/server/library/tv-episode-resolver.js';
 
 interface SourceMediaFile {
 	path: string;
